extract bundle_meta
