+++ bundle/src/lib.rs
@@ -1,9 +1,11 @@
+mod bundle_meta;
 mod bundler;
 mod custom_tag;
 mod files;
 mod types;
 
-pub use bundler::BundlerUtil;
+pub use bundle_meta::*;
+pub use bundler::*;
 pub use custom_tag::*;
 pub use files::*;
 pub use types::*;
